requirements/base.txt
Django==5.1.*
psycopg[binary]==3.*
celery[redis]==5.*
django-environ==0.*
httpx==0.*
djangorestframework==3.*
djangorestframework-simplejwt==5.*
